Pass ExtCap if metadata but fails if none
Various changes to try to get Extended cap element present when set
metadata. Not sure which important but finally found I was throwing
error in encodeExtendedOperations method rather than just doing nothing
so this stopped anything being created. Fixing this now means the
testNoInspireElementWhenNoMetadata fails as element is generated;
it just failed before because of bug!