Remove unused version check code
This actually caused an error on deployed version which doesn't occur
in tests. Somehow the test environment did set the version whereas it
was null in the deployed environment. Can't write a test for this as I
don't know how these set up and, as not used, not worth it for the
moment.